Update to compile all templates on startup, add html.ResonseWriter to Pipeline, ...
The template compilation solves a problem with having dynamic components always compiling templates.

This means we renmaed .Compile() to .Init().

We now let the user assign the name, so they can manage collisions. This is way better than the name+uuid thing I was doing.  Must have been tired when I came up with that.

The Element type got split.  Now it just requires Execute() and is user implementable.

DyanmicFunc() takes Pipeline as an argument. Infinitely more powerful.

compileElement(s) is now used to use reflection to call Init() on any Element or []Elmement that has it.

Added iframe support.