Massive WASM update
This has a lot of things that I'm sure broke testa and compatibility.

But I'm putting up as we are exp anyways.

This supports wasm and components working together.  We now have an updater for controlling changes base on events.  We support event handling and event listeners.

So much cleanup to do, but it works, HOLY CRAP!!!